Creted login component
